requirements.txt
grpcio==1.64.1
grpcio-reflection==1.64.1
grpcio-tools==1.64.1
mypy-protobuf==3.6.0
protobuf==5.27.1
types-protobuf==5.26.0.20240422
